+++ src/main/java/school/sorokin/eventmanager/locations/dto/ResponseLocationDto.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public class ResponseLocationDto {
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -20,7 +20,7 @@ public class ResponseLocationDto {
 
     }
 
-    public ResponseLocationDto(Integer id, String name, String address, Integer capacity, String description) {
+    public ResponseLocationDto(Long id, String name, String address, Integer capacity, String description) {
         this.id = id;
         this.name = name;
         this.address = address;
@@ -28,11 +28,11 @@ public class ResponseLocationDto {
         this.description = description;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
